backend/requirements.txt
fastapi==0.111.1
uvicorn==0.32.0
sqlalchemy==2.0.36
mysql-connector-python==8.4.0
pydantic==2.9.2
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
aiosqlite==0.19.0
redis==5.0.1
aioredis==2.0.1
python-dotenv==1.0.0
loguru==0.7.2
python-json-logger==2.0.7
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-sqlalchemy==0.47b0